move checkBox listener to new module
