+++ src/modules/functions.js
@@ -160,13 +160,3 @@ export const addTaskToHTML = (task, arrayOfTasks) => {
     editTask(event.target.parentNode, arrayOfTasks);
   });
 };
-
-export const addCheckboxListenerOnLoad = () => {
-  const checkboxes = document.querySelectorAll('input[type=checkbox]');
-  checkboxes.forEach((checkbox) => {
-    checkbox.addEventListener('change', () => {
-      const textTask = checkbox.nextElementSibling;
-      textTask.classList.toggle('line-through');
-    });
-  });
-};
